Revert to export from index
Avoid dealing with dependency stubbing issue
